requirements.txt
tensorflow>=2.0
opencv-python
pillow
